fix env variables managment
