+++ nucliadb_node/binding/src/lib.rs
@@ -21,7 +21,7 @@ use std::io::Cursor;
 
 use nucliadb_core::paragraphs::ParagraphIterator;
 use nucliadb_core::texts::DocumentIterator;
-use nucliadb_node::config::Configuration;
+use nucliadb_node::env;
 use nucliadb_node::reader::NodeReaderService as RustReaderService;
 use nucliadb_node::writer::NodeWriterService as RustWriterService;
 use nucliadb_protos::{
@@ -583,7 +583,7 @@ impl NodeWriter {
 
 #[pymodule]
 fn nucliadb_node_binding(_py: Python, m: &PyModule) -> PyResult<()> {
-    let log_levels = Configuration::log_level();
+    let log_levels = env::log_level();
 
     let mut layers = Vec::new();
     let stdout_layer = tracing_subscriber::fmt::layer()
